api/requirements.txt
# FastAPI y servidor
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Transcripción de audio - solo faster-whisper que es más confiable
faster-whisper>=0.10.0

# Procesamiento de audio
ffmpeg-python>=0.2.0
pydub>=0.25.1

# Utilidades
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Validación y serialización
pydantic>=2.0.0

# Logging y monitoreo
loguru>=0.7.0

# Testing y WebSocket client (opcional para testing)
websockets>=11.0.0
aiohttp>=3.8.0
